backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.7.1
pymongo>=4.9,<4.12
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
xgboost>=2.1.3
joblib==1.4.2
openpyxl==3.1.5
python-dotenv==1.0.1
openai==1.59.7
Pillow==11.1.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
torchvision==0.20.1+cpu
transformers==4.47.1
httpx==0.28.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
Faker==33.3.1
